SM-1287: Refactor out completions and config command from main.rs in bws
