+++ crates/bws/src/command/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod completions;
+pub(crate) mod config;
